+++ Scripts/WoW Classic Game Data API/Item API/Endpoint_WoWClassic_GameData_ItemClass.cs
@@ -1,0 +1,74 @@
+// ╔════════════════════════════════════╗
+// ║ This file has been auto-generated. ║
+// ╚════════════════════════════════════╝
+
+using System;
+using System.Collections;
+using ZeShmouttsAssets.BlizzardAPI.JSON;
+
+namespace ZeShmouttsAssets.BlizzardAPI
+{
+	/// <summary>
+	/// Interface for working with the Blizzard API inside Unity.
+	/// </summary>
+	public static partial class BlizzardAPI
+	{
+		/// <summary>
+		/// API endpoints related to World of Warcraft Classic game data (items, spells, etc.).
+		/// Reference : https://develop.battle.net/documentation/world-of-warcraft-classic/game-data-apis
+		/// </summary>
+		public static partial class WowClassicGameData
+		{
+			internal const string apiPath_ItemClass = basePath_Wow_gameData + "/item-class/{0}";
+
+			/// <summary>
+			/// Coroutine that retrieves an item class by ID.
+			/// </summary>
+			/// <param name="itemClassId">The ID of the item class.</param>
+			/// <param name="action_Result">Action to execute with the data once retrieved and converted.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetItemClass(int itemClassId, Action<Json_WowClassic_ItemClass> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = string.Format(apiPath_ItemClass, itemClassId);
+				yield return SendRequest(region, namespaceClassicStatic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+			/// <summary>
+			/// Coroutine that retrieves an item class by ID, as a raw JSON string.
+			/// </summary>
+			/// <param name="itemClassId">The ID of the item class.</param>
+			/// <param name="action_Result">Action to execute with the raw JSON string.</param>
+			/// <param name="ifModifiedSince">Adds a request header to check if the document has been modified since this date (in HTML format), which will return an empty response body if it's older.</param>
+			/// <param name="action_LastModified">Action to execute with the date of the last server-side modification to the document.</param>
+			/// <param name="region">The region of the data to retrieve.</param>
+			/// <returns></returns>
+			public static IEnumerator GetItemClassRaw(int itemClassId, Action<string> action_Result, string ifModifiedSince = null, Action<string> action_LastModified = null, BattleNetRegion region = DefaultRegion)
+			{
+				string path = string.Format(apiPath_ItemClass, itemClassId);
+				yield return SendRequest(region, namespaceClassicStatic, path, action_Result, ifModifiedSince, action_LastModified);
+			}
+
+		}
+	}
+}
+
+namespace ZeShmouttsAssets.BlizzardAPI.JSON
+{
+	/// <summary>
+	/// JSON structure for World of Warcraft (Classic), representing an item class.
+	/// </summary>
+	[Serializable]
+	public class Json_WowClassic_ItemClass : Object_JSON
+	{
+		// {{JSON_START}}
+		public LinkStruct _links;
+
+		public int class_id;
+		public LocalizedString name;
+		public RefNameIdStruct[] item_subclasses;
+		// {{JSON_END}}
+	}
+}